requirements.txt
absl-py==0.8.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bert-serving-client==1.10.0
bert-serving-server==1.10.0
blis==0.4.1
boto==2.49.0
boto3==1.12.26
botocore==1.15.26
catalogue==1.0.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.2.2
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
cytoolz==0.10.1
dask==2.8.1
decorator==4.4.0
docutils==0.15.2
filelock==3.0.12
flickrapi==2.4.0
gast==0.3.2
gensim==3.8.0
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.23.0
h5py==2.10.0
httplib2==0.17.3
idna==2.9
imageio==2.6.1
importlib-metadata==1.5.0
-e git+https://github.com/pwdyson/inflect.py@863b51997ff8dc1b76480c74f8e85e128d98355a#egg=inflect
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
leveldb==0.20
Markdown==3.1.1
matplotlib==3.1.1
murmurhash==1.0.0
networkx==2.4
nltk==3.4.5
numpy==1.17.3
oauth2==1.9.0.post1
oauthlib==3.1.0
olefile==0.46
packaging==20.8
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.3.0
plac==0.9.6
preshed==3.0.2
progress==1.5
prompt-toolkit==2.0.9
protobuf==3.11.1
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.4.2
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==19.0.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
sacremoses==0.0.38
scikit-image==0.14.2
scikit-learn==0.21.3
scipy==1.3.2
sentence-transformers==0.3.9
sentencepiece==0.1.91
simplenlg==0.2.0
six==1.12.0
smart-open==1.8.4
spacy==2.2.3
srsly==1.0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.6.0
tensorflow-probability==0.7.0
termcolor==1.1.0
thinc==7.3.0
tokenizers==0.9.3
toolz==0.10.0
torch==1.7.1
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.2
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.25.7
virtualenv==16.7.7
wasabi==0.6.0
wcwidth==0.1.7
Werkzeug==0.16.0
wget==3.2
wrapt==1.11.2
zipp==3.1.0
